+++ src/test/java/org/springframework/beans/factory/HelloService.java
@@ -1,5 +1,7 @@
 package org.springframework.beans.factory;
 
+import lombok.Setter;
+
 /**
  * @author: DoubleW2w
  * @date: 2024/9/29
@@ -8,16 +10,20 @@ package org.springframework.beans.factory;
 public class HelloService {
 
   private String name;
+  @Setter private WorldService worldService;
 
   public HelloService(String name) {
     this.name = name;
   }
 
-  public HelloService() {
-  }
+  public HelloService() {}
 
   public String sayHello() {
-    System.out.println("hello");
+    System.out.println("name=" + name);
     return "hello";
+  }
+
+  public void sayWorld() {
+    worldService.world();
   }
 }